feat(v2/3): Allow handle v2 and v3 requests

+++ _requests/delete/unadopt.ts
@@ -1,11 +1,21 @@
+import { User } from "@supabase/supabase-js";
 import { VercelRequest, VercelResponse } from "@vercel/node";
+import { urlContainsV3 } from "../../_utils/check-if-v3";
 import { supabase } from "../../_utils/supabase";
 
 export default async function handler(
 	request: VercelRequest,
-	response: VercelResponse
+	response: VercelResponse,
+	user?: User
 ) {
-	const { tree_id, uuid } = request.body;
+	let { uuid } = request.body;
+	const { tree_id } = request.body;
+	if (!request.url) {
+		return response.status(500).json({ error: "no url in request" });
+	}
+	if (urlContainsV3(request.url)) {
+		uuid = user?.id || uuid;
+	}
 	const { error } = await supabase
 		.from("trees_adopted")
 		.delete()
